+++ CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.12](https://github.com/@nativescript-community/ui-collectionview/compare/v5.3.11...v5.3.12) (2024-01-14)
+
+### Bug Fixes
+
+* prevent requestLayout in Akylas fork while updating cell in svelte ([4a1239b](https://github.com/@nativescript-community/ui-collectionview/commit/4a1239b5160f2b60e802c099a1fca301e9ff5014))
+
+## [5.3.11](https://github.com/@nativescript-community/ui-collectionview/compare/v5.3.10...v5.3.11) (2023-12-08)
+
+### Bug Fixes
+
+* notify of disposing on `disposeNativeView` so that flavours can destroy their instances ([98bf75e](https://github.com/@nativescript-community/ui-collectionview/commit/98bf75e9db15ba0c4e30431f828917b8b2c52a98))
+
+## [5.3.10](https://github.com/@nativescript-community/ui-collectionview/compare/v5.3.9...v5.3.10) (2023-12-06)
+
+### Bug Fixes
+
+* **collectionview:** `itemOverlap` fixes ([a473aa9](https://github.com/@nativescript-community/ui-collectionview/commit/a473aa96e5fcc5529081bfc1048141f672a223a2))
+* **collectionview:** bring back `displayItem` event ([0c73f3c](https://github.com/@nativescript-community/ui-collectionview/commit/0c73f3c85c0435641b931330b718dc19b4a259ed))
+* **collectionview:** native api usage fix ([d0e544b](https://github.com/@nativescript-community/ui-collectionview/commit/d0e544bf2f651213f1a3c6a30cabe0cc48a0c2ca))
+* **swipemenu:** `swipeMenuId` on the collectionView to allow the swipeMenu to not be the root child of the template ([ab4addb](https://github.com/@nativescript-community/ui-collectionview/commit/ab4addb0bbac8be0de0b0cd53aa209e8d4085e12))
+
 ## [5.3.9](https://github.com/@nativescript-community/ui-collectionview/compare/v5.3.8...v5.3.9) (2023-11-30)
 
 **Note:** Version bump only for package ui-collectionview